# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Identity Platform API",
    "version": "1.0.0",
    "description": "Wire-level API contract for the Identity Platform. The semantic guarantees for these endpoints live in docs/contracts."
  },
  "servers": [
    {
      "url": "/",
      "description": "This deployment"
    },
    {
      "url": "https://auth.example.com",
      "description": "Production example"
    },
    {
      "url": "http://localhost:5300",
      "description": "Local development"
    }
  ],
...
  "paths": {
...
    "/auth/email/verify": {
      "post": {
        "tags": ["Account lifecycle"],
        "summary": "Verify an email with an emailed token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/EmailVerifyRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Email verified",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MessageResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
    "/auth/verify": {
      "post": {
        "tags": ["Verification"],
        "summary": "Verify a JWT or API key remotely",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/VerifyRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Verification result",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/VerifyResponse" }
              }
            }
          },
          "400": {
            "description": "The expected audience or credential was not provided",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/VerifyResponse" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Use the form: ApiKey sk_..."
      },
      "adminKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Admin-Key"
      }
    },
...
    "responses": {
      "ValidationError": {
        "description": "Validation error",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
...
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": { "type": "string" },
          "code": { "type": "string" },
          "details": { "type": "array", "items": {} }
        }
      },
      "MessageResponse": {
        "type": "object",
        "required": ["message"],
        "properties": {
          "message": { "type": "string" }
        }
      },
...
      "ClientCredentialFields": {
        "type": "object",
        "required": ["clientId"],
        "properties": {
          "clientId": { "type": "string", "example": "cl_example" },
          "clientSecret": { "type": "string", "example": "cs_example" }
        }
      },
...
      "EmailVerifyRequest": {
        "allOf": [
          { "$ref": "#/components/schemas/ClientCredentialFields" },
          {
            "type": "object",
            "required": ["token"],
            "properties": {
              "token": { "type": "string" }
            }
          }
        ]
      },
...
      "VerifyRequest": {
        "type": "object",
        "required": ["audience"],
        "oneOf": [
          { "required": ["token"] },
          { "required": ["apiKey"] }
        ],
        "properties": {
          "token": { "type": "string" },
          "apiKey": { "type": "string" },
          "audience": {
            "type": "string",
            "description": "External client ID of the application accepting the credential",
            "example": "cl_0123456789abcdef"
          },
          "requiredPermission": { "type": "string", "example": "users:read" }
        }
      },
      "VerifyResponse": {
        "type": "object",
        "required": ["valid"],
        "properties": {
          "valid": { "type": "boolean" },
          "authorized": { "type": "boolean" },
          "error": { "type": "string" },
          "user": {
            "type": "object",
            "properties": {
              "id": { "type": "string", "format": "uuid" },
              "clientId": { "type": "string", "format": "uuid" },
              "email": { "type": "string", "format": "email" },
              "permissions": { "type": "array", "items": { "type": "string" } }
            }
          },
          "apiKey": {
            "type": "object",
            "properties": {
              "clientId": { "type": "string", "format": "uuid" },
              "name": { "type": "string" },
              "scopes": { "type": "array", "items": { "type": "string" } },
              "serviceAccount": {
                "type": "object",
                "properties": {
                  "id": { "type": "string", "format": "uuid" },
                  "name": { "type": "string" }
                }
              }
            }
          }
        }
      },
...
    }
  }
}